Refactor Bandcamp content integration: move music filter to app module and update manifest
