+++ src/bandcamp/content/app.ts
@@ -1,13 +1,19 @@
-import { mount } from 'svelte';
+import { mount, unmount } from 'svelte';
 import App from './app.svelte';
 import './app.css';
 import { getExtensionUrl } from 'src/utils/chrome.runtime';
 import { injectCSSFile } from 'src/utils/dom';
+import 'src/utils/console';
+import BCXMusicFilter from '$lib/components/bcx/BCXMusicFilter.svelte';
+import { getMusicItems } from '../page/music/html';
 
 console.log('Bandcamp content app module!');
 
+let musicFilterComponent: any = null;
+
 function init() {
   mountApp();
+  mountMusicFilter();
 }
 
 function mountApp() {
@@ -32,6 +38,47 @@ function mountApp() {
   );
 }
 
+function mountMusicFilter() {
+  if (!window.location.hostname.includes('bandcamp.com')) {
+    return;
+  }
+
+  const musicItems = getMusicItems();
+
+  if (musicItems.length === 0) {
+    console.log('No music items found on this page');
+    return;
+  }
+
+  // Create a container for the music filter component
+  const musicGrid = document.getElementById('music-grid');
+  if (!musicGrid) {
+    console.log('Music grid not found on this page');
+    return;
+  }
+
+  // Create container element for the Svelte component
+  const filterContainer = document.createElement('div');
+  filterContainer.id = 'bcx-music-filter';
+
+  // Insert the container before the music grid
+  if (musicGrid.parentNode) {
+    musicGrid.parentNode.insertBefore(filterContainer, musicGrid);
+  }
+
+  // Mount the Svelte component using Svelte 5 syntax
+  musicFilterComponent = mount(BCXMusicFilter, {
+    target: filterContainer,
+    props: {
+      musicItems: musicItems,
+    },
+  });
+
+  console.log(
+    `BCX Music Filter component mounted with ${musicItems.length} items`,
+  );
+}
+
 if (document.readyState === 'loading') {
   document.addEventListener('DOMContentLoaded', init);
 } else {
@@ -42,3 +89,11 @@ if (document.readyState === 'loading') {
 if (document.readyState === 'complete') {
   setTimeout(init, 100);
 }
+
+// Cleanup music filter component on page unload
+window.addEventListener('beforeunload', () => {
+  if (musicFilterComponent) {
+    unmount(musicFilterComponent);
+    musicFilterComponent = null;
+  }
+});
